+++ hydromt_wflow/wflow_sbm.py
@@ -1120,7 +1120,6 @@ setting new flood_depth dimensions"
         * **glacier_initial_leq_depth** map: storage (volume) of glacier per cell [mm]
 
         Required setup methods:
-
         * :py:meth:`~WflowSbmModel.setup_basemaps`
 
         Parameters
@@ -1915,6 +1914,7 @@ Run setup_soilmaps first"
         self,
         soil_fn: str = "soilgrids",
         ptf_ksatver: str = "brakensiek",
+        ptf_hb: str = "brakensiek",
         wflow_thicknesslayers: list[int] = [100, 300, 800],
         output_names: dict = {
             "soil_water__saturated_volume_fraction": "soil_theta_s",
@@ -1923,6 +1923,7 @@ Run setup_soilmaps first"
             "soil__thickness": "soil_thickness",
             "soil_water__vertical_saturated_hydraulic_conductivity_scale_parameter": "soil_f",  # noqa: E501
             "soil_layer_water__brooks_corey_exponent": "soil_brooks_corey_c",
+            "soil_water__air_entry_pressure_head": "soil_hb",
         },
     ):
         """
@@ -1934,8 +1935,9 @@ Run setup_soilmaps first"
 
         Currently, supported ``soil_fn`` is "soilgrids" and "soilgrids_2020".
         ``ptf_ksatver`` (PTF for the vertical hydraulic conductivity) options are
-        "brakensiek" and "cosby". "soilgrids" provides data at 7 specific depths,
-        while "soilgrids_2020" provides data averaged over 6 depth intervals.
+        "brakensiek" and "cosby". ``ptf_hb`` (PTF for the air entry pressure)
+        options are "brakensiek" and "clapp". "soilgrids" provides data at 7 specific
+        depths, while "soilgrids_2020" provides data averaged over 6 depth intervals.
         This leads to small changes in the workflow:
         (1) M parameter uses midpoint depths in soilgrids_2020 versus \
 specific depths in soilgrids,
@@ -1978,6 +1980,9 @@ a map for each of the wflow_sbm soil layers (n in total)
         * **meta_soil_texture** map: soil texture based on USDA soil texture triangle \
 (mapping: [1:Clay, 2:Silty Clay, 3:Silty Clay-Loam, 4:Sandy Clay, 5:Sandy Clay-Loam, \
 6:Clay-Loam, 7:Silt, 8:Silt-Loam, 9:Loam, 10:Sand, 11: Loamy Sand, 12:Sandy Loam])
+        * **soil_water__air_entry_pressure_head** map:
+            The air entry pressure [cm] at which air begins to enter the largest \
+            pores of a saturated soil during drying.
 
 
         Required setup methods:
@@ -1998,6 +2003,10 @@ a map for each of the wflow_sbm soil layers (n in total)
             Pedotransfer function (PTF) to use for calculation of ksat vertical
             (vertical saturated hydraulic conductivity [mm/day]).
             By default 'brakensiek'.
+        ptf_hb : {'brakensiek', 'clapp'}
+            Pedotransfer function (PTF) method to use for calculation of hb
+            (air entry pressure [cm]).
+            By default 'brakensiek'.
         wflow_thicknesslayers : list of int, optional
             Thickness of soil layers [mm] for wflow_sbm soil model.
             By default [100, 300, 800] for layers at depths 100, 400, 1200 and >1200 mm.
@@ -2016,6 +2025,7 @@ a map for each of the wflow_sbm soil layers (n in total)
             ds=dsin,
             ds_like=self.staticmaps.data,
             ptfKsatVer=ptf_ksatver,
+            ptf_hb=ptf_hb,
             soil_fn=soil_fn,
             wflow_layers=wflow_thicknesslayers,
         ).reset_coords(drop=True)